requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
#apex==0.1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bokeh==3.1.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.5
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
Django==4.2
enum34==1.1.10
filelock==3.10.7
flatbuffers==23.5.26
fonttools==4.37.3
frozenlist==1.3.3
#future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google==3.0.0
google-auth==1.35.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gorilla==0.4.0
grpcio==1.40.0
h5py==3.9.0
hyperopt==0.2.7
idna==3.2
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.17.3
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
Markdown==3.3.4
MarkupSafe==2.1.3
matplotlib==3.6.0
msgpack==1.0.5
multidict==6.0.4
munch==2.5.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
#nvidia-dali-cuda110==1.16.1
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==2.0.0
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
platformdirs==3.2.0
#pretrainedmodels==0.7.4
prettytable==3.7.0
protobuf==4.24.0
psutil==5.9.4
py-spy==0.3.14
py4j==0.10.9.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
pyrsistent==0.19.3
#pystopwatch2==0.1.1
python-dateutil==2.8.2
pytorch-optimizer==2.11.0
pytz==2023.3
PyYAML==6.0
ray==0.8.4
redis==4.5.4
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.2.2
scipy==1.10.0
seaborn==0.12.2
six==1.16.0
#sklearn==0.0.post1
smmap==5.0.0
soupsieve==2.4
sqlparse==0.4.3
tabulate==0.9.0
#tensorboard==2.13.0
#tensorboard-data-server==0.7.1
#tensorboard-plugin-wit==1.8.0
#tensorboardX==2.6
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
#theconf==0.1.0
threadpoolctl==3.1.0
timm==0.6.7
#torch==1.8.1+cu111
#torchaudio==0.8.1
torchinfo==1.8.0
torchsummary==1.5.1
#torchvision==0.9.1+cu111
tornado==6.3.2
tqdm==4.65.0
typing-extensions==3.10.0.2
tzdata==2023.3
urllib3==1.26.6
virtualenv==20.21.0
#visdom==0.2.4
#warmup-scheduler==0.1
wcwidth==0.2.6
websocket-client==1.6.1
Werkzeug==2.0.1
wrapt==1.15.0
xyzservices==2023.5.0
yarl==1.8.2
zipp==3.15.0
